refactor: Extract color_class module.
This breaks some indirect dependencies on subs
and message_view_header.
