+++ static/js/stream_list.js
@@ -5,6 +5,7 @@ import render_stream_privacy from "../templates/stream_privacy.hbs";
 import render_stream_sidebar_row from "../templates/stream_sidebar_row.hbs";
 
 import * as blueslip from "./blueslip";
+import * as color_class from "./color_class";
 import * as hash_util from "./hash_util";
 import * as keydown_util from "./keydown_util";
 import {ListCursor} from "./list_cursor";
@@ -13,7 +14,6 @@ import * as narrow_state from "./narrow_state";
 import * as popovers from "./popovers";
 import * as resize from "./resize";
 import * as scroll_util from "./scroll_util";
-import * as stream_color from "./stream_color";
 import * as stream_data from "./stream_data";
 import * as stream_popover from "./stream_popover";
 import * as stream_sort from "./stream_sort";
@@ -254,7 +254,7 @@ function build_stream_sidebar_li(sub) {
         color: sub.color,
         pin_to_top: sub.pin_to_top,
     };
-    args.dark_background = stream_color.get_color_class(args.color);
+    args.dark_background = color_class.get_css_class(args.color);
     const list_item = $(render_stream_sidebar_row(args));
     return list_item;
 }
@@ -311,7 +311,7 @@ export function redraw_stream_privacy(sub) {
     }
 
     const div = li.find(".stream-privacy");
-    const dark_background = stream_color.get_color_class(sub.color);
+    const dark_background = color_class.get_css_class(sub.color);
 
     const args = {
         invite_only: sub.invite_only,
